+++ __init__.py
@@ -1,5 +1,5 @@
 #-*- coding: utf-8 -*-
 from utility import parseFloat, parseInt, optimizeValue
-from logics import Interpreter, JSCompiler #, Parser        (will fail on PyJS at run-time)
+from logics import Interpreter  #, Parser        (will fail on PyJS at run-time)
 from parser import ParseException
 import vistache